pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crunch-cli"
version = "0.1.0"
description = "A pipe-friendly CLI tool for tabular data transformations"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["csv", "data", "cli", "transform", "pipeline", "tabular"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "typer[all]>=0.12",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "tabulate>=0.9",
    "scikit-learn>=1.4",
    "numpy>=1.26",
]

[project.optional-dependencies]
plot    = ["altair>=5.2", "vl-convert-python>=1.0", "plotext>=5.2", "scipy>=1.12"]
parquet = ["pyarrow>=14.0"]
all     = ["crunch-cli[plot,parquet]"]

[project.scripts]
crunch = "crunch.main:app"

[project.urls]
Homepage = "https://github.com/yourname/crunch-cli"
Issues   = "https://github.com/yourname/crunch-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/crunch"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas-stubs>=2.0",
]

[tool.hatch.envs.default.scripts]
test      = "pytest {args:tests} --cov=src/crunch --cov-report=term-missing"
lint      = "ruff check src tests"
typecheck = "mypy src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
timeout = 30
addopts  = "-v"
